Split the Project module into smaller modules and use defdelegate
