+++ lib/sanbase/model/project/project.ex
@@ -1,14 +1,14 @@
 defmodule Sanbase.Model.Project do
   use Ecto.Schema
   import Ecto.Changeset
+
+  alias __MODULE__
   alias Sanbase.Repo
 
   alias Sanbase.Model.{
-    Project,
     ProjectEthAddress,
     ProjectBtcAddress,
     Ico,
-    IcoCurrency,
     Currency,
     MarketSegment,
     Infrastructure,
@@ -99,190 +99,28 @@ defmodule Sanbase.Model.Project do
     |> unique_constraint(:coinmarketcap_id)
   end
 
-  @spec describe(%Project{}) :: String.t()
-  def describe(%Project{coinmarketcap_id: cmc_id}) when not is_nil(cmc_id) do
-    "project with slug: #{cmc_id}"
-  end
+  defdelegate roi_usd(project), to: Project.Roi
 
-  def describe(%Project{id: id}) do
-    "project with id: #{id}"
-  end
+  defdelegate initial_ico(project), to: Project.FundsRaised
+
+  defdelegate funds_raised_usd_ico_end_price(project), to: Project.FundsRaised
+
+  defdelegate funds_raised_eth_ico_end_price(project), to: Project.FundsRaised
+
+  defdelegate funds_raised_btc_ico_end_price(project), to: Project.FundsRaised
+
+  defdelegate funds_raised_icos(project), to: Project.FundsRaised
+
+  defdelegate describe(project), to: Project.Description
+
+  defdelegate contract_info_by_slug(slug), to: Project.ContractData
+
+  defdelegate contract_info(project), to: Project.ContractData
+
+  defdelegate contract_address(project), to: Project.ContractData
 
   def sanbase_link(%Project{coinmarketcap_id: cmc_id}) when not is_nil(cmc_id) do
     SanbaseWeb.Endpoint.frontend_url() <> "/projects/#{cmc_id}"
-  end
-
-  def initial_ico(%Project{id: id}) do
-    Ico
-    |> where([i], i.project_id == ^id)
-    |> first(:start_date)
-    |> Repo.one()
-  end
-
-  @doc ~S"""
-  ROI = current_price*(ico1_tokens + ico2_tokens + ...)/(ico1_tokens*ico1_initial_price + ico2_tokens*ico2_initial_price + ...)
-  We skip ICOs for which we can't calculate the initial_price or the tokens sold
-  For ICOs that we don't have tokens sold we try to fill it heuristically by evenly distributing the rest of the total available supply
-  """
-  def roi_usd(%Project{ticker: ticker, coinmarketcap_id: coinmarketcap_id} = project)
-      when not is_nil(ticker) and not is_nil(coinmarketcap_id) do
-    with %Project{} = project <- Repo.preload(project, [:latest_coinmarketcap_data, :icos]),
-         false <- is_nil(project.latest_coinmarketcap_data),
-         false <- is_nil(project.latest_coinmarketcap_data.price_usd),
-         false <- is_nil(project.latest_coinmarketcap_data.available_supply) do
-      zero = Decimal.new(0)
-
-      tokens_and_initial_prices =
-        project
-        |> fill_missing_tokens_sold_at_icos()
-        |> Enum.map(fn ico ->
-          {ico.tokens_sold_at_ico, calc_token_usd_ico_price_by_project(project, ico)}
-        end)
-        |> Enum.reject(fn {tokens_sold_at_ico, token_usd_ico_price} ->
-          is_nil(tokens_sold_at_ico) or is_nil(token_usd_ico_price)
-        end)
-
-      total_cost =
-        tokens_and_initial_prices
-        |> Enum.map(fn {tokens_sold_at_ico, token_usd_ico_price} ->
-          Decimal.mult(tokens_sold_at_ico, token_usd_ico_price)
-        end)
-        |> Enum.reduce(zero, &Decimal.add/2)
-
-      total_gain =
-        tokens_and_initial_prices
-        |> Enum.map(fn {tokens_sold_at_ico, _} -> tokens_sold_at_ico end)
-        |> Enum.reduce(zero, &Decimal.add/2)
-        |> Decimal.mult(project.latest_coinmarketcap_data.price_usd)
-
-      case total_cost do
-        ^zero -> nil
-        total_cost -> Decimal.div(total_gain, total_cost)
-      end
-    else
-      _ -> nil
-    end
-  end
-
-  def roi_usd(_), do: nil
-
-  # Private functions
-
-  # Heuristic: fills empty ico.tokens_sold_at_ico by evenly distributing the rest of the circulating supply
-  # TODO:
-  # Currently uses latest_coinmarketcap_data.available_supply, which also includes coins not issued at any ICO
-  # Maybe it's better to keep historical data of available_supply so that we can calculate it better
-  defp fill_missing_tokens_sold_at_icos(%Project{} = project) do
-    with tokens_sold_at_icos <- Enum.map(project.icos, & &1.tokens_sold_at_ico),
-         unknown_count <- Enum.filter(tokens_sold_at_icos, &is_nil/1) |> length(),
-         true <- unknown_count > 0 do
-      zero = Decimal.new(0)
-      one = Decimal.new(1)
-
-      known_tokens_sum =
-        tokens_sold_at_icos
-        |> Enum.reject(&is_nil/1)
-        |> Enum.reduce(zero, &Decimal.add/2)
-
-      unknown_tokens_sum =
-        Decimal.compare(project.latest_coinmarketcap_data.available_supply, known_tokens_sum)
-        |> case do
-          ^one ->
-            Decimal.sub(project.latest_coinmarketcap_data.available_supply, known_tokens_sum)
-
-          _ ->
-            zero
-        end
-
-      unknown_tokens_single_ico = Decimal.div(unknown_tokens_sum, Decimal.new(unknown_count))
-
-      Enum.map(project.icos, fn ico ->
-        if is_nil(ico.tokens_sold_at_ico) do
-          Map.put(ico, :tokens_sold_at_ico, unknown_tokens_single_ico)
-        else
-          ico
-        end
-      end)
-    else
-      _ -> project.icos
-    end
-  end
-
-  defp calc_token_usd_ico_price_by_project(%Project{} = project, %Ico{} = ico) do
-    ico.token_usd_ico_price ||
-      calc_token_usd_ico_price(
-        ico.token_eth_ico_price,
-        "ETH",
-        ico.start_date,
-        project.latest_coinmarketcap_data.update_time
-      ) ||
-      calc_token_usd_ico_price(
-        ico.token_btc_ico_price,
-        "BTC",
-        ico.start_date,
-        project.latest_coinmarketcap_data.update_time
-      )
-  end
-
-  defp calc_token_usd_ico_price(nil, _currency_from, _ico_start_date, _current_datetime), do: nil
-  defp calc_token_usd_ico_price(_price_from, _currency_from, nil, _current_datetime), do: nil
-
-  defp calc_token_usd_ico_price(price_from, currency_from, ico_start_date, current_datetime) do
-    with :gt <- Ecto.DateTime.compare(current_datetime, Ecto.DateTime.from_date(ico_start_date)),
-         timestamp <- Sanbase.DateTimeUtils.ecto_date_to_datetime(ico_start_date),
-         price_usd when not is_nil(price_usd) <-
-           Sanbase.Prices.Utils.fetch_last_price_before(currency_from, "USD", timestamp) do
-      Decimal.mult(price_from, Decimal.new(price_usd))
-    else
-      _ -> nil
-    end
-  end
-
-  def funds_raised_usd_ico_end_price(project) do
-    funds_raised_ico_end_price(project, &Ico.funds_raised_usd_ico_end_price/1)
-  end
-
-  def funds_raised_eth_ico_end_price(project) do
-    funds_raised_ico_end_price(project, &Ico.funds_raised_eth_ico_end_price/1)
-  end
-
-  def funds_raised_btc_ico_end_price(project) do
-    funds_raised_ico_end_price(project, &Ico.funds_raised_btc_ico_end_price/1)
-  end
-
-  defp funds_raised_ico_end_price(project, ico_funds_raised_fun) do
-    Repo.preload(project, :icos).icos
-    |> Enum.map(ico_funds_raised_fun)
-    |> Enum.reject(&is_nil/1)
-    |> case do
-      [] -> nil
-      amounts -> Enum.reduce(amounts, 0, &Kernel.+/2)
-    end
-  end
-
-  @doc """
-  For every currency aggregates all amounts for every ICO of the given project
-  """
-  def funds_raised_icos(%Project{id: id}) do
-    query =
-      from(
-        i in Ico,
-        inner_join: ic in IcoCurrency,
-        on: ic.ico_id == i.id and not is_nil(ic.amount),
-        inner_join: c in Currency,
-        on: c.id == ic.currency_id,
-        where: i.project_id == ^id,
-        group_by: c.code,
-        order_by: fragment("case
-                            			when ? = 'BTC' then '_'
-                            			when ? = 'ETH' then '__'
-                            			when ? = 'USD' then '___'
-                            		  else ?
-                            		end", c.code, c.code, c.code, c.code),
-        select: %{currency_code: c.code, amount: sum(ic.amount)}
-      )
-
-    Repo.all(query)
   end
 
   def supply(%Project{} = project) do
@@ -311,89 +149,33 @@ defmodule Sanbase.Model.Project do
     |> Repo.one()
   end
 
-  def by_slug(slug) when is_binary(slug) do
+  def by_slug(slug, opts \\ [])
+
+  def by_slug(slug, opts) when is_binary(slug) do
+    additional_preloads = Keyword.get(opts, :additional_preloads, [])
+
     Project
     |> where([p], p.coinmarketcap_id == ^slug)
-    |> preload(^@preloads)
+    |> preload(^(@preloads ++ additional_preloads))
     |> Repo.one()
   end
 
-  def by_slug(slugs) when is_list(slugs) do
+  def by_slug(slugs, opts) when is_list(slugs) do
+    additional_preloads = Keyword.get(opts, :additional_preloads, [])
+
     Project
     |> where([p], p.coinmarketcap_id in ^slugs)
-    |> preload(^@preloads)
+    |> preload(^(@preloads ++ additional_preloads))
     |> Repo.all()
   end
 
-  def by_id(id) when is_integer(id) or is_binary(id) do
+  def by_id(id, opts \\ []) when is_integer(id) or is_binary(id) do
+    additional_preloads = Keyword.get(opts, :additional_preloads, [])
+
     Project
     |> where([p], p.id == ^id)
-    |> preload(^@preloads)
+    |> preload(^(@preloads ++ additional_preloads))
     |> Repo.one()
-  end
-
-  def contract_info_by_slug("ethereum"), do: {:ok, "ETH", 18}
-  def contract_info_by_slug("bitcoin"), do: {:ok, "BTC", 8}
-
-  def contract_info_by_slug(slug) do
-    from(p in Project,
-      where: p.coinmarketcap_id == ^slug,
-      select: {p.main_contract_address, p.token_decimals}
-    )
-    |> Repo.one()
-    |> case do
-      {contract, token_decimals} when is_binary(contract) ->
-        {:ok, String.downcase(contract), token_decimals || 0}
-
-      _ ->
-        {:error, {:missing_contract, "Can't find contract address of project with slug: #{slug}"}}
-    end
-  end
-
-  def contract_address(%Project{} = project) do
-    case contract_info(project) do
-      {:ok, address, _} -> address
-      _ -> nil
-    end
-  end
-
-  # Internally when we have a table with blockchain related data
-  # contract address is used to identify projects. In case of ethereum
-  # the contract address contains simply 'ETH'
-  def contract_info(%Project{coinmarketcap_id: "ethereum"}), do: {:ok, "ETH", 18}
-  def contract_info(%Project{coinmarketcap_id: "bitcoin"}), do: {:ok, "BTC", 8}
-
-  def contract_info(%Project{
-        main_contract_address: main_contract_address,
-        token_decimals: token_decimals
-      })
-      when not is_nil(main_contract_address) do
-    {:ok, String.downcase(main_contract_address), token_decimals || 0}
-  end
-
-  def contract_info(%Project{} = project) do
-    {:error, {:missing_contract, "Can't find contract address of #{describe(project)}"}}
-  end
-
-  def contract_info(data) do
-    {:error, "Not valid project type provided to contract_info - #{inspect(data)}"}
-  end
-
-  def eth_addresses_by_tickers(tickers) do
-    query =
-      from(
-        p in Project,
-        where: p.ticker in ^tickers and not is_nil(p.coinmarketcap_id),
-        preload: [:eth_addresses]
-      )
-
-    Repo.all(query)
-    |> Stream.map(fn %Project{ticker: ticker, eth_addresses: eth_addresses} ->
-      eth_addresses = eth_addresses |> Enum.map(&Map.get(&1, :address))
-
-      {ticker, eth_addresses}
-    end)
-    |> Enum.into(%{})
   end
 
   def ticker_by_slug(nil), do: nil
@@ -461,6 +243,23 @@ defmodule Sanbase.Model.Project do
     {:ok, addresses}
   end
 
+  def eth_addresses_by_tickers(tickers) do
+    query =
+      from(
+        p in Project,
+        where: p.ticker in ^tickers and not is_nil(p.coinmarketcap_id),
+        preload: [:eth_addresses]
+      )
+
+    Repo.all(query)
+    |> Stream.map(fn %Project{ticker: ticker, eth_addresses: eth_addresses} ->
+      eth_addresses = eth_addresses |> Enum.map(&Map.get(&1, :address))
+
+      {ticker, eth_addresses}
+    end)
+    |> Enum.into(%{})
+  end
+
   @doc """
   Return all projects from the list which trading volume is over a given threshold
   """
@@ -509,10 +308,17 @@ defmodule Sanbase.Model.Project do
     parse_github_organization_link(github_link, slug)
   end
 
-  def is_erc20?()
+  def is_erc20?(%Project{} = project) do
+    project
+    |> Repo.preload(:infrastructure)
+  end
 
   def preloads(), do: @preloads
-  def preload(items), do: Repo.preload(items, @preloads)
+
+  def preload_assocs(items, opts \\ []) do
+    additional_preloads = Keyword.get(opts, :additional_preloads, [])
+    Repo.preload(items, @preloads ++ additional_preloads)
+  end
 
   defp parse_github_organization_link(github_link, slug) do
     # nil will break the regex
